refactor : TeamList 컴포넌트 이벤트 함수 분리
- 코드의 통일성을 위해 TeamList 컴포넌트 이벤트 함수 분리
