+++ src/js/components/TeamList.js
@@ -1,10 +1,9 @@
 import { $ } from "../lib/util.js";
 import { TEMPLATE } from "../constants/template.js";
-import { fetchRequest } from "../lib/fetchRequest.js";
-import { API_URL, METHOD } from "../constants/config.js";
-import { INFORM_MESSAGES, ERROR_MESSAGES } from "../constants/message.js";
 
 import TeamModel from "./model/TeamModel.js";
+
+import { addTeam } from "./event/TeamList.js";
 
 class TeamList {
   constructor({ teamListData }) {
@@ -31,20 +30,8 @@ class TeamList {
     $(".team-list-container").innerHTML = template.join("") + TEMPLATE.ADD_TEAM_BUTTON;
   }
 
-  async addTeam() {
-    const teamName = prompt(INFORM_MESSAGES.ADD_TEAM);
-    if (!teamName) return;
-
-    const { response, error } = await fetchRequest(API_URL.TEAMS, METHOD.POST, { name: teamName });
-
-    if (error) return alert(ERROR_MESSAGES.ADD_TEAM);
-
-    this.teamListData.push(new TeamModel({ ...response, id: response._id }));
-    this.render();
-  }
-
   registerEventListener() {
-    $("#add-team-button").addEventListener("click", this.addTeam.bind(this));
+    $("#add-team-button").addEventListener("click", addTeam.bind(this));
   }
 }
